+++ tests/test_structure.py
@@ -1,4 +1,8 @@
-from strainer.structure import field, create_serializer, child, many
+import pytest
+
+from strainer.structure import field, dict_field, create_serializer, child, many
+from strainer import validators
+from strainer.exceptions import ValidationException
 
 
 class ChildTestObject(object):
@@ -28,6 +32,21 @@ def test_field():
     assert {'a': 1} == target
 
 
+def test_dict_field():
+    serializer = dict_field('a')
+    test_obj = {'a': 'b'}
+    target = {}
+    serializer.to_representation(test_obj, target)
+
+    assert {'a': 'b'} == target
+
+    from_json = {'a': 'b'}
+    target = {}
+    serializer.to_internal(from_json, target)
+
+    assert {'a': 'b'} == target
+
+
 def test_field_custom_to_representation():
     serializer = field('a', to_representation=lambda x: x.a + 1)
     test_obj = TestObject()
@@ -50,6 +69,21 @@ def test_field_multiple():
     serializer.to_internal(from_json, target)
 
     assert {'d': [1, 2]} == target
+
+
+def test_field_multiple_validation():
+    serializer = field('e', multiple=True, validators=[validators.string(max_length=1)])
+    test_obj = {
+      'e': ['a', 'bb']
+    }
+    target = {}
+    errors = None
+    try:
+        serializer.to_internal(test_obj, target)
+    except ValidationException as e:
+        errors = e.errors
+
+    assert errors == {'e': {1: ['This field is to long, max length is 1']}}
 
 
 def test_serializer():
@@ -125,3 +159,47 @@ def test_many():
     assert reference == target
     target = serializer.to_internal(reference)
     assert target == reference
+
+
+def test_nested_required():
+    child_serializer = create_serializer(
+        field('c1', validators=[validators.required()])
+    )
+
+    serializer = create_serializer(
+        field('a'),
+        many('c', serializer=child_serializer, validators=[validators.required()])
+    )
+
+    reference = {
+      'a': 1,
+    }
+
+    with pytest.raises(ValidationException):
+        serializer.to_internal(reference)
+
+    reference = {
+      'a': 1,
+      'c': [],
+    }
+
+    errors = None
+    try:
+        serializer.to_internal(reference)
+    except ValidationException as e:
+        errors = e.errors
+
+    assert errors == {'c': ['This field is required']}
+
+    reference = {
+      'a': 1,
+      'c': [{}],
+    }
+
+    errors = None
+    try:
+        serializer.to_internal(reference)
+    except ValidationException as e:
+        errors = e.errors
+
+    assert errors == {'c': [{'c1': ['This field is required']}]}